refactor: #60 피드백 반영 #55
- 엔티티 디렉토리 제거로 통일
- 예외처리 부정 반전
- 중복된 deletedAt 조건 제거
- dto 내 변수 네이밍 수정
